Move DspControl to oddio, as it is oddio specific
DspControl has a lifetime,
therefore it will not work with bevy_audio and kira
